Update and rename Day_1 - Calculator(GUI) to Day_1 - WATER CONSUMPTION
The provided Java code reads an integer 'n' from the console and then proceeds to read 'n' more integers. For each of these input integers, if the value is greater than or equal to 2000, it prints "YES"; otherwise, it prints "NO". Essentially, it determines whether each input integer is greater than or equal to 2000 and provides a corresponding "YES" or "NO" output.